+++ tests/test_web_api.py
@@ -3,38 +3,68 @@
 """
 A module that contains methods for testing the web_api module
 """
+
 import re
+from unittest import mock
 
 from pytest import raises
 
 from masquerade.providers.web_api import (
-    WEB_API_URL,
+    BASE_URL,
     etl_candidate,
     get_candidate_from_parts,
     get_candidate_from_single_line,
     get_candidates_from_single_line,
+    reverse_geocode,
 )
 
 
-def test_etl_candidate():
+@mock.patch("masquerade.providers.open_sgid.get_city")
+def test_etl_candidate(get_city_mock):
+    get_city_mock.return_value = "WEST VALLEY CITY"
     input = {
-        "address": "3989 SOUTH FRONTAGE RD, Salt Lake City",
+        "address": "3989 SOUTH FRONTAGE RD, SALT LAKE CITY",
         "location": {"x": 416707.62244415266, "y": 4508458.9846219225},
         "score": 69.92,
         "locator": "Centerlines.StatewideRoads",
-        "addressGrid": "Salt Lake City",
+        "addressGrid": "SALT LAKE CITY",
     }
 
-    result = etl_candidate(input)
+    result = etl_candidate(input, 26912)
 
-    assert result["address"] == input["address"]
+    assert result["address"] == "3989 SOUTH FRONTAGE RD, WEST VALLEY CITY"
     assert result["attributes"]["score"] == 69.92
     assert result["attributes"]["locator"] == input["locator"]
     assert result["attributes"]["addressGrid"] == input["addressGrid"]
     assert result["location"] == input["location"]
 
 
-def test_etl_candidate_base_result():
+@mock.patch("masquerade.providers.open_sgid.get_city")
+@mock.patch("masquerade.providers.open_sgid.get_county")
+def test_etl_candidate_fallback_to_county(get_county_mock, get_city_mock):
+    get_city_mock.return_value = None
+    get_county_mock.return_value = "SALT LAKE"
+    input = {
+        "address": "3989 SOUTH FRONTAGE RD, SALT LAKE CITY",
+        "location": {"x": 416707.62244415266, "y": 4508458.9846219225},
+        "score": 69.92,
+        "locator": "Centerlines.StatewideRoads",
+        "addressGrid": "SALT LAKE CITY",
+    }
+
+    result = etl_candidate(input, 26912)
+
+    assert result["address"] == "3989 SOUTH FRONTAGE RD, SALT LAKE COUNTY"
+    assert result["attributes"]["score"] == 69.92
+    assert result["attributes"]["locator"] == input["locator"]
+    assert result["attributes"]["addressGrid"] == input["addressGrid"]
+    assert result["location"] == input["location"]
+
+
+@mock.patch("masquerade.providers.open_sgid.get_city")
+def test_etl_candidate_base_result(get_city_mock):
+    get_city_mock.return_value = "HOLLADAY"
+
     #: this is a test for the base result object as opposed to result.candidate objects
     input = {
         "location": {"x": 429340.24421129236, "y": 4504146.207401402},
@@ -47,9 +77,9 @@ def test_etl_candidate_base_result():
         "candidates": [],
     }
 
-    result = etl_candidate(input)
+    result = etl_candidate(input, 26912)
 
-    assert result["address"] == input["matchAddress"]
+    assert result["address"] == "3987 S 1925 E, HOLLADAY"
     assert result["score"] == 100
     assert result["attributes"]["locator"] == input["locator"]
     assert result["attributes"]["addressGrid"] == input["addressGrid"]
@@ -107,7 +137,7 @@ def test_get_address_candidate(requests_mock):
         "status": 200,
     }
 
-    requests_mock.get(re.compile(f"{WEB_API_URL}.*"), json=mock_response)
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode.*"), json=mock_response)
     candidates = get_candidates_from_single_line("123 s main, 84115", 3857, 5)
 
     assert len(candidates) == 5
@@ -164,7 +194,7 @@ def test_get_address_candidate_perfect_match(requests_mock):
         "status": 200,
     }
 
-    requests_mock.get(re.compile(f"{WEB_API_URL}.*"), json=mock_response)
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode.*"), json=mock_response)
     candidates = get_candidates_from_single_line("123 s main st, 84115", 3857, 5)
 
     assert len(candidates) == 1
@@ -185,7 +215,7 @@ def test_get_address_candidate_single_result_for_batch(requests_mock):
         "status": 200,
     }
 
-    requests_mock.get(re.compile(f"{WEB_API_URL}.*"), json=mock_response)
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode.*"), json=mock_response)
     candidate = get_candidate_from_parts("123 s main st", "84115", 3857)
 
     assert candidate["score"] == 80
@@ -205,7 +235,7 @@ def test_get_address_candidate_no_candidates(requests_mock):
         "status": 200,
     }
 
-    requests_mock.get(re.compile(f"{WEB_API_URL}.*"), json=mock_response)
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode.*"), json=mock_response)
     candidate = get_candidate_from_parts("123 s main st", "84115", 3857)
 
     assert candidate is None
@@ -225,14 +255,14 @@ def test_get_candidate_from_single_line(requests_mock):
         "status": 200,
     }
 
-    requests_mock.get(re.compile(f"{WEB_API_URL}.*"), json=mock_response)
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode.*"), json=mock_response)
     candidate = get_candidate_from_single_line("123 s main st, 84115", 3857)
 
     assert candidate["score"] == 100
 
 
 def test_get_address_candidates_raises(requests_mock):
-    requests_mock.get(re.compile(f"{WEB_API_URL}.*"), json={}, status_code=500)
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode.*"), json={}, status_code=500)
 
     with raises(Exception):
         get_candidates_from_single_line("123 s main street, 84114", 3857, 5)
@@ -240,7 +270,7 @@ def test_get_address_candidates_raises(requests_mock):
 
 def test_get_address_candidates_bad_address(requests_mock):
     requests_mock.get(
-        re.compile(f"{WEB_API_URL}.*"),
+        re.compile(f"{BASE_URL}/geocode.*"),
         json={
             "status": 404,
             "message": "No address candidates found with a score of 70 or better.",
@@ -255,3 +285,113 @@ def test_get_address_candidates_bad_address(requests_mock):
     candidates = get_candidates_from_single_line("0101300036", 3857, 5)
 
     assert len(candidates) == 0
+
+
+@mock.patch("masquerade.providers.open_sgid.get_city")
+@mock.patch("masquerade.providers.open_sgid.get_county")
+@mock.patch("masquerade.providers.open_sgid.get_zip")
+def test_reverse_geocode(get_zip_mock, get_county_mock, get_city_mock, requests_mock):
+    get_city_mock.return_value = "SALT LAKE CITY"
+    get_county_mock.return_value = "SALT LAKE"
+    get_zip_mock.return_value = "84101"
+
+    mock_response = {
+        "result": {
+            "address": {
+                "street": "123 S MAIN ST",
+                "addressType": "PointAddress",
+            }
+        },
+        "status": 200,
+    }
+
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode/reverse.*"), json=mock_response)
+
+    result = reverse_geocode(-12455627.277556794, 4977968.997941715, 3857, -12455627.277556794, 4977968.997941715)
+
+    assert result["Match_addr"] == "123 S MAIN ST, SALT LAKE CITY, UTAH, 84101"
+    assert result["LongLabel"] == "123 S MAIN ST, SALT LAKE CITY, UTAH, 84101, USA"
+    assert result["ShortLabel"] == "123 S MAIN ST"
+    assert result["Addr_type"] == "PointAddress"
+    assert result["AddNum"] == "123"
+    assert result["Address"] == "123 S MAIN ST"
+    assert result["City"] == "SALT LAKE CITY"
+    assert result["Subregion"] == "SALT LAKE COUNTY"
+    assert result["Region"] == "UTAH"
+    assert result["Postal"] == "84101"
+    assert result["CntryName"] == "UNITED STATES"
+    assert result["CountryCode"] == "USA"
+    assert result["X"] == -12455627.277556794
+    assert result["Y"] == 4977968.997941715
+    assert result["InputX"] == -12455627.277556794
+    assert result["InputY"] == 4977968.997941715
+
+
+@mock.patch("masquerade.providers.open_sgid.get_city")
+@mock.patch("masquerade.providers.open_sgid.get_county")
+@mock.patch("masquerade.providers.open_sgid.get_zip")
+def test_reverse_geocode_no_city(get_zip_mock, get_county_mock, get_city_mock, requests_mock):
+    get_city_mock.return_value = None
+    get_county_mock.return_value = "SALT LAKE"
+    get_zip_mock.return_value = "84101"
+
+    mock_response = {
+        "result": {
+            "address": {
+                "street": "123 S MAIN ST",
+                "addressType": "PointAddress",
+            }
+        },
+        "status": 200,
+    }
+
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode/reverse.*"), json=mock_response)
+
+    result = reverse_geocode(-12455627.277556794, 4977968.997941715, 3857, -12455627.277556794, 4977968.997941715)
+
+    assert result["Match_addr"] == "123 S MAIN ST, SALT LAKE COUNTY, UTAH, 84101"
+    assert result["LongLabel"] == "123 S MAIN ST, SALT LAKE COUNTY, UTAH, 84101, USA"
+    assert result["ShortLabel"] == "123 S MAIN ST"
+    assert result["Addr_type"] == "PointAddress"
+    assert result["AddNum"] == "123"
+    assert result["Address"] == "123 S MAIN ST"
+    assert result["City"] == ""
+    assert result["Subregion"] == "SALT LAKE COUNTY"
+    assert result["Region"] == "UTAH"
+    assert result["Postal"] == "84101"
+    assert result["CntryName"] == "UNITED STATES"
+    assert result["CountryCode"] == "USA"
+    assert result["X"] == -12455627.277556794
+    assert result["Y"] == 4977968.997941715
+    assert result["InputX"] == -12455627.277556794
+    assert result["InputY"] == 4977968.997941715
+
+
+@mock.patch("masquerade.providers.open_sgid.get_city")
+@mock.patch("masquerade.providers.open_sgid.get_county")
+@mock.patch("masquerade.providers.open_sgid.get_zip")
+def test_reverse_geocode_no_response(get_zip_mock, get_county_mock, get_city_mock, requests_mock):
+    get_city_mock.return_value = "SALT LAKE CITY"
+    get_county_mock.return_value = "SALT LAKE"
+    get_zip_mock.return_value = "84101"
+
+    requests_mock.get(re.compile(f"{BASE_URL}/geocode/reverse.*"), status_code=500)
+
+    result = reverse_geocode(-12455627.277556794, 4977968.997941715, 3857, -12455627.277556794, 4977968.997941715)
+
+    assert result["Match_addr"] == ""
+    assert result["LongLabel"] == ""
+    assert result["ShortLabel"] == ""
+    assert result["Addr_type"] == ""
+    assert result["AddNum"] == ""
+    assert result["Address"] == ""
+    assert result["City"] == "SALT LAKE CITY"
+    assert result["Subregion"] == "SALT LAKE COUNTY"
+    assert result["Region"] == "UTAH"
+    assert result["Postal"] == "84101"
+    assert result["CntryName"] == "UNITED STATES"
+    assert result["CountryCode"] == "USA"
+    assert result["X"] == -12455627.277556794
+    assert result["Y"] == 4977968.997941715
+    assert result["InputX"] == -12455627.277556794
+    assert result["InputY"] == 4977968.997941715